+++ src/cascadia/TerminalSettingsEditor/CompatibilityViewModel.idl
